fixtures: split apps into apps and servers
