+++ mir-ci/mir_ci/tests/test_drag_and_drop.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from textwrap import dedent
 
 import pytest
-from mir_ci.fixtures import apps
+from mir_ci.fixtures import servers
+from mir_ci.program.app import App
 from mir_ci.program.display_server import DisplayServer
 from mir_ci.wayland.screencopy_tracker import ScreencopyTracker
 from mir_ci.wayland.virtual_pointer import VirtualPointer
@@ -49,11 +50,11 @@ ${{END_TEMPLATE}}    {TESTS_PATH}/robot/templates/drag_and_drop_end.png
 @pytest.mark.parametrize(
     "modern_server",
     [
-        apps.ubuntu_frame(),
-        # apps.mir_kiosk(), we need servers based on Mir 2.14 or later
-        apps.confined_shell(),
-        apps.mir_test_tools(),
-        apps.mir_demo_server(),
+        servers.ubuntu_frame(),
+        # servers.mir_kiosk(), we need servers based on Mir 2.14 or later
+        servers.confined_shell(),
+        servers.mir_test_tools(),
+        servers.mir_demo_server(),
     ],
 )
 @pytest.mark.deps(
@@ -81,7 +82,7 @@ class TestDragAndDrop:
     async def test_source_and_dest_match(self, modern_server, app, tmp_path) -> None:
         extensions = VirtualPointer.required_extensions + ScreencopyTracker.required_extensions
         server_instance = DisplayServer(modern_server, add_extensions=extensions)
-        program = server_instance.program(apps.App(app))
+        program = server_instance.program(App(app))
 
         robot_test_case = dedent(
             """\
@@ -97,7 +98,7 @@ class TestDragAndDrop:
             robot_file.write(
                 ROBOT_TEMPLATE.format(settings=ROBOT_SETTINGS, variables=ROBOT_VARIABLES, test_case=robot_test_case)
             )
-            robot = server_instance.program(apps.App(("robot", "-d", tmp_path, robot_file.name)))
+            robot = server_instance.program(App(("robot", "-d", tmp_path, robot_file.name)))
 
             async with server_instance, program, robot:
                 await robot.wait(60)
@@ -115,7 +116,7 @@ class TestDragAndDrop:
     async def test_source_and_dest_mismatch(self, modern_server, app, tmp_path) -> None:
         extensions = VirtualPointer.required_extensions + ScreencopyTracker.required_extensions
         server_instance = DisplayServer(modern_server, add_extensions=extensions)
-        program = server_instance.program(apps.App(app))
+        program = server_instance.program(App(app))
 
         robot_test_case = dedent(
             """\
@@ -132,7 +133,7 @@ class TestDragAndDrop:
             robot_file.write(
                 ROBOT_TEMPLATE.format(settings=ROBOT_SETTINGS, variables=ROBOT_VARIABLES, test_case=robot_test_case)
             )
-            robot = server_instance.program(apps.App(("robot", "-d", tmp_path, robot_file.name)))
+            robot = server_instance.program(App(("robot", "-d", tmp_path, robot_file.name)))
 
             async with server_instance, program, robot:
                 await robot.wait(60)
